# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Analizcim API",
    "description": "Excel analiz uygulaması için REST API. Bu API, satış ve alış dosyalarını analiz etmek, geçmişi yönetmek, tahminler oluşturmak ve kullanıcı yönetimi için kullanılır.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@analizcim.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Yerel geliştirme sunucusu"
    }
  ],
...
  "paths": {
...
    "/history": {
      "get": {
        "tags": ["Geçmiş"],
        "summary": "Analiz geçmişini listele",
        "description": "Tüm analiz kayıtlarını sayfalamalı olarak listeler.",
        "operationId": "getHistory",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Filtreleme için yıl"
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Arama terimi"
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["date_asc", "date_desc", "id_asc", "id_desc"]
            },
            "description": "Sıralama"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Sayfa başına kayıt sayısı"
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Kaydın başlangıç konumu"
          }
        ],
        "responses": {
          "200": {
            "description": "Geçmiş listesi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "history": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "date": {
                            "type": "string"
                          },
                          "displayDate": {
                            "type": "string"
                          },
                          "salesFileName": {
                            "type": "string"
                          },
                          "purchaseFileName": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Geçmiş"],
        "summary": "Tüm geçmişi temizle",
        "description": "Tüm analiz geçmişini siler.",
        "operationId": "clearHistory",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Geçmiş temizlendi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/history/{id}": {
      "get": {
        "tags": ["Geçmiş"],
        "summary": "Belirli bir geçmiş kaydını getir",
        "operationId": "getHistoryById",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Geçmiş kaydı ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Geçmiş kaydı",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "entry": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Kayıt bulunamadı",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Geçmiş"],
        "summary": "Belirli bir geçmiş kaydını sil",
        "operationId": "deleteHistoryById",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Geçmiş kaydı ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Kayıt silindi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Kayıt bulunamadı"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Oturum kimlik doğrulaması"
      },
      "adminAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Admin yetkisi gerektiren işlemler için"
      }
    }
  }
}